Moved kinect dependent tests to new files
