+++ src/test/java/edu/mines/acmX/exhibit/input_services/hardware/HardwareManagerTest.java
@@ -214,59 +214,4 @@ public class HardwareManagerTest {
 		hm.resetAllDrivers();		
 	}
 	
-	@Test
-	public void testValidDriverCache()
-			throws HardwareManagerManifestException, BadDeviceFunctionalityRequestException,
-			InvalidConfigurationFileException, BadFunctionalityRequestException {
-		log.info("testValidDriverConfigFile");
-		
-		HardwareManager.setManifestFilepath(BASE_FILE + "GoodCompleteManifest.xml");
-		HardwareManager hm = HardwareManager.getInstance();
-
-		Map<String, DependencyType> mmd = new HashMap<String, DependencyType>();
-		hm.setRunningModulePermissions(mmd);
-		hm.resetAllDrivers();
-		
-		mmd.put("depth", DependencyType.REQUIRED);
-		hm.setRunningModulePermissions(mmd);
-
-		Map<String, String> configStore = new HashMap<String, String>();
-		configStore.put("kinectopenni", "openni_config.xml");
-		
-		hm.setConfigurationFileStore(configStore);
-		hm.resetAllDrivers();
-		
-		assertTrue(hm.getNumberofDriversInCache() == 1);
-	}
-	
-	@Test
-	public void testOptionalFunctionalityGetsLoadedAtRuntime()
-			throws HardwareManagerManifestException,
-			InvalidConfigurationFileException,
-			BadFunctionalityRequestException {
-		
-		log.info("testOptionalFunctionalityGetsLoadedAtRuntime");
-		
-		HardwareManager.setManifestFilepath(BASE_FILE + "GoodCompleteManifest.xml");
-		HardwareManager hm = HardwareManager.getInstance();
-
-		Map<String, DependencyType> mmd = new HashMap<String, DependencyType>();
-		hm.setRunningModulePermissions(mmd);
-		hm.resetAllDrivers();
-		
-		mmd.put("depth", DependencyType.OPTIONAL);
-		hm.setRunningModulePermissions(mmd);
-
-		Map<String, String> configStore = new HashMap<String, String>();
-		configStore.put("kinectopenni", "openni_config.xml");
-		
-		hm.setConfigurationFileStore(configStore);
-		hm.resetAllDrivers();
-		
-		assertTrue(hm.getNumberofDriversInCache() == 0);
-		List<String> devices = hm.getDevices("depth");
-		assertTrue(hm.getNumberofDriversInCache() == 1);
-	}
-	
-	
 }
